requirements.txt
# WorthIt! Lite dependencies
python-telegram-bot>=20.6
fastapi>=0.104.1
uvicorn>=0.24.0
httpx>=0.24.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0
anyio>=4.1.0
starlette>=0.27.0
apify-client>=1.5.1
pydantic>=2.5.2
python-multipart>=0.0.6
h2>=4.1.0  # HTTP/2 support for httpx
redis>=5.0.1  # Redis Cloud support
tenacity>=8.2.3  # For retry logic
slowapi>=0.1.8  # Rate limiting support
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.21.0
python-dotenv>=1.0.0
httpx>=0.24.0
redis>=5.0.0
tenacity>=8.2.2

# ML and data processing
nltk>=3.8.1
textblob>=0.17.1
numpy>=1.24.3
opencv-python>=4.7.0

# Monitoring and metrics
prometheus-client>=0.16.0
prometheus-fastapi-instrumentator>=6.0.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0

# Telegram bot
python-telegram-bot>=20.3

# Web scraping
requests>=2.28.2

# API rate limiting
slowapi>=0.1.7

# Async support
aiohttp>=3.8.4

# Security
python-multipart>=0.0.6

# Payment processing
stripe>=11.6.0
